Moved CI test partition creation into batch script.
Something changed on Travis' Windows VMs within the last 3 days that
caused our cmd.exe /C invocation to stop working (and to just drop into
a shell). This commit attempts to remedy the situation by relocating
this command into a batch script and invoking that instead (which Bash
on Windows is supposed to understand a bit more intuitively).

Signed-off-by: Jacob Howard <jacob@havoc.io>