Refactored data channel connection wrapper to use read loop.
We've had to revert the data channel connection wrapper to use a read
loop, although this time with significantly better state management. It
turns out that data channels won't yield error and close events when
detached, because they rely on a read loop for that. Thus, it made sense
to reinstate our read loop.

Signed-off-by: Jacob Howard <jacob@havoc.io>