Forced usage of older Windows container image on Travis.
It seems like the KB4532691 tag of the 1809 image, released on
February 11, 2020, is stalling out container builds. Other users are
also reporting this online. This commit reverts to the KB4534273 tag,
released on January 14, 2020, in attempt to avoid these breaking
changes.

We should revert this eventually, but not until a new image build is
provided.

Signed-off-by: Jacob Howard <jacob@havoc.io>