Added Windows named pipe support and cleaned up UDS listening code.
To better support Docker client forwarding on Windows, this commit adds
support for forwarding Windows named pipes. This support is minimal for
the time being, with no configuration options and permissions that
default to creator-only access. This should be fine for the vast
majority of use cases, and there's nothing that prevents additional
configuration options in the future.

This commit also cleans up the logic around Unix domain socket listener
creation, finally taking advantage of Go 1.13's support for error
unwrapping to accurately classify errors. It also cleans up support for
Unix domain sockets on Windows. This probably should have been a
separate commit, but the existing code was sort of in the way.

This new named pipe support still stays within minor version
compatibility, which only matters for tunnel hosting anyway. If the
"npipe" protocol is specified for forwarding with an agent that doesn't
support it, then the session initialization will simply fail with an
error. Tunnel agents can be updated if this is a concern.

Signed-off-by: Jacob Howard <jacob@havoc.io>