Cleaned up comment in Docker container probing.
This commit was originally going to be focused on adding standard error
output to the combined error messaged generated when probing fails, but
the potential error output (~32K) is too much to include in an error
string, even if it's unlikely to be that long. However, the comment
surrounding this error handling needed a bit of a cleanup, so this
commit takes care of that.

Signed-off-by: Jacob Howard <jacob@havoc.io>