Added lazy initialization support to listener endpoints.
This commit modifies local listener endpoints to support lazy
initialization on Open. This prevents issues with "phantom" listeners
that don't actually accept connections while the forwarding loop is
still trying to connect to the destination. This was causing issues with
Docker commands (which don't have an internal HTTP timeout) that would
never exit because they had an HTTP request stuck on a connection from a
listener's accept backlog (which was never actually accepted because the
forwarding loop couldn't connect to the remote destination).

At the moment, we only enable this lazy initialization for local
endpoints since (a) those are the only endpoints where we can reliably
report lazy initialization failures and (b) the local-to-remote
forwarding scenario is where we saw this issue crop up. Additionally,
the remote-to-local forwarding scenario doesn't have this problem since
the local destination endpoint will always connect immediately and
without issue because it doesn't require a network connection and
doesn't need to to establish a listener.

The one scenario not addressed by this fix is remote-to-remote
forwarding, but (a) this isn't currently a widely used and (b) the
"phantom" listener issue hasn't been observed or reported to cause
problems in these cases. If we do want to enable lazy initialization for
remote listeners, it will require a refactor of the forwarding protocol.
Given the unclear benefit, we'll hold off on that for now.

Signed-off-by: Jacob Howard <jacob@havoc.io>