Reduced dialing asynchrony and disabled lazy listening for tests.
This commit reduces asynchronous endpoint dialing to only tunnel-based
endpoints. This dialing was already synchronous for SSH and Docker
endpoints (so there's no change there), so only local endpoints are
affected by this change. Since local endpoints already connect
immediately, there's no benefit to making their connection operations
asynchronous. In fact, making them connect synchronously makes it easier
to report their errors and avoid session creation if something fails.

This commit also adds an internal flag to disable lazy listening for
integration tests. Combined with synchronous local endpoint
connectivity, this change allows us to simplify the forwarding
integration test significantly. The one downside is that this is a bit
of a hack and slightly misaligns the test environment with reality, but
the test still covers the forwarding internals. Given that lazy
listening is merely a heuristic and that the previous test structure was
even further from real usage, this seems like an improvement.

Signed-off-by: Jacob Howard <jacob@havoc.io>