Refactored compose command scaffolding.
This commit refactors the way that command dispatch takes place in the
compose command. We now explicitly scaffold out all of the Docker
Compose command surface and rely on more traditional Cobra command
dispatch (with a few hacks to enable handling of unknown commands). The
splicing of the Mutagen CLI and Docker Compose isn't exactly perfect,
but it's probably as close as it can be given their drastically
different command line parsing libraries. The full scaffolding, while
tedious, is probably going to be necessary in most cases to facilitate
the hooks that we need. It will also be essential to enable shell
completion (without writing some really ugly hacks).

The current functionality is minimal, but the help and version commands
are implemented and should present some idea of how to directly shell
out to Docker Compose with unmodified arguments.

A flexible utility function for shelling out to Docker Compose has also
been implemented.

Signed-off-by: Jacob Howard <jacob@havoc.io>