Added emulation of Docker Compose project loading.
In order to facilitate inspection and injection of services with our
Docker Compose wrapper, we need to emulate computation of configuration
files, project working directory, and project name. This logic is
somewhat non-trivial, so we try to emulate Docker Compose's behavior as
faithfully as possible. The only difference is that we work with
absolute paths rather than relative paths, both for efficiency and
robustness.

Signed-off-by: Jacob Howard <jacob@havoc.io>