Implemented most compose commands as pass-throughs.
This commit implements most of the compose commands (all except for
up/down, pause/unpause, and stop/start/restart) as pass-through
commands. For the pass-through commands that it implements, it adds flag
definitions to enable shell completion. Additional work can be done on
shell completion to bring in some of the value detection mechanisms that
Docker Compose's shell completion supports. If we're clever and
interface directly with the Docker API at some point, then we may be
able to outperform Docker Compose completion. If we write this using
Cobra, it should also port more readily to other shells where Docker
Compose completion support isn't as extensive.

Signed-off-by: Jacob Howard <jacob@havoc.io>