Cleaned up Docker Compose invocation function.
This commit replaces the compose function with a simpler version that
takes fewer parameters. It also updates the passthrough handler and
other calls to use this new version. The original version was designed
for invoking Docker Compose in a flexible manner, but realistically
anything that interacts with Docker Compose in a more complex fashion is
going to need custom separate code.

Signed-off-by: Jacob Howard <jacob@havoc.io>