Implemented Docker Compose configuration loading and interpolation.
This commit implements loading and interpolation for Docker Compose
configuration files. This required using v3 of the YAML parsing package
to access raw YAML nodes for interpolation. Unfortuantely it wasn't
possible to simply invoke docker-compose config because it doesn't print
top-level extension fields (which we need to grab x-mutagen
information). It also wasn't possible to use the compose-go package's
YAML loading/interpolation because it uses the v2 YAML package
internally and wouldn't allow us to interpolate numeric types.

The one downside to this approach is that it doesn't provide the strict
YAML validation provided by the v2 YAML package. However, in a future
commit, we may be able to re-serialize the YAML after interpolation and
then re-parse it using the v2 package. This can be avoided later once
go-yaml/yaml#460 is resolved (potentially via go-yaml/yaml#557).

This commit also adds the license notice that we'll need for the
compose-go package.

Signed-off-by: Jacob Howard <jacob@havoc.io>