Refactored compose YAML loading mechanism.
This commit refactors the compose command's YAML loading mechanism to
enable strict decoding for the x-mutagen section of the YAML file. This
requires a slight hack (namely re-serializing the interpolated YAML),
but it seems to work fairly reliably. It also allows us to continue
using non-strict decoding for other elements of the Docker Compose YAML.

Signed-off-by: Jacob Howard <jacob@havoc.io>