Added parameters field to URLs.
This commit adds a parameters field to URLs that can be used to store
internal parameters for a transport. The exact nature of these
parameters is undefined and transport-dependent. They'll be needed to
support Docker URLs that take into account command line parameters. They
may also be useful later for passing command line options to OpenSSH and
other transport commands.

Signed-off-by: Jacob Howard <jacob@havoc.io>