Relocated Docker Compose project loading to compose package.
This commit relocates the Docker Compose project loading emulation code
to the compose package. This code was originally living alongside the
command implementation because it seemed too specific to that command to
put in the generic compose package, but it seems like it has a better
chance of being cleaned up and made more generic if it lives in the
separate compose package.

Signed-off-by: Jacob Howard <jacob@havoc.io>