Comment cleanup in Docker Compose project loading.
This commit cleans up some of the comments in the compose package,
particularly those dealing with project loading. The project loading
function is still more complex than I'd like, but I feel like it's been
boiled down to its essential complexity, and given that it has to
emulate the behavior of a third-party tool, I'm not sure what else can
be done to simplify it. I've tried extracting portions of its code into
reusable routines, but their behavior is so contrived that it's less
confusing to simply inline the code. I think we'll just have to live
with this for the time being.

Signed-off-by: Jacob Howard <jacob@havoc.io>