Updated Mutagen service to use pre-built image in compose package.
This commit updates Mutagen's Docker Compose integration to use a
pre-built image for the Mutagen service. This will be more reliable and
extensible than the ad hoc image generation strategy that we were using.

The ad hoc strategy had a few problems. The biggest problem was the
build context path changing every run, which caused Docker Compose to
rebuild the service image and hence the service container every run.
This was because the com.docker.compose.config-hash label, which takes
into account the build context path, was changing every time. Second,
when trying to ensure that the Mutagen service was up, it was necessary
to know whether or not the service image was already built, because
specifying --build would lead to image/container recreation but omitting
it would lead to a warning. Third, the entry point we were using was
hacky and didn't have good signal handling behavior. Fourth, the build
was slower than just pulling an image.

This new strategy will allow us to make the Mutagen service smarter
(though it's still just a no-op at the moment).

Signed-off-by: Jacob Howard <jacob@havoc.io>