Added URL comparison to session reconciliation in compose up.
This commit adds an Equal comparison method to URL and then uses that to
extend the session comparison performed during session reconciliation in
compose up.

It still feels like a more nuanced approach might be necessary for cases
where URL parameters might change regularly and still point to the same
daemon, but I think the more complicated we make this comparison, the
more likely it is to break. In terms of behaving in-line with Docker
Compose, I think that our use of the project name and daemon identifier
effectively emulates the same level of isolation. Any URL comparisons
that fail due to parameters changing regularly are something that the
user can fix/avoid.

Signed-off-by: Jacob Howard <jacob@havoc.io>