Added a warning to compose run passthrough.
While we still implement run as a passthrough command, the truth is that
it doesn't interact well with Mutagen due to its disregard for
network/volume creation and service startup ordering. We continue to
implement it as a passthrough, but we've added a warning to its
invocation. Users can use docker-compose directly to avoid this warning.

We've also updated the Warning function to use colorized output so that
its output doesn't get lost visually.

Signed-off-by: Jacob Howard <jacob@havoc.io>