Revert "Enabled command name replacement in compose help."
This reverts commit b833d6cc90cfbadb17bfcced69a6164ef9cded7c.

This replacement wasn't particularly helpful and there wasn't a clean
way to extend it to passthrough commands (where we disable flag parsing
and don't intercept help requests). The implementation is worth
preserving in a revert, and the stream package extraction still makes
sense since we may use that code elsewhere too.