Added new dependencies and compatible software to license information.
This commit adds Docker Compose as a dependency in the license
information. Although we don't depend on Docker Compose directly, we do
incorporate ported logic from its codebase and thus should indicate a
dependency of some sort.

This commit also adds a list of software with which Mutagen is
compatible, namely OpenSSH, Docker, and Docker Compose. Since we don't
depend on or distribute these elements, we've excluded their licensing
information.

Signed-off-by: Jacob Howard <jacob@havoc.io>