Decoupled prompting from services.
This commit is the first half of an effort to decouple prompting from
the forwarding, synchronization, and tunneling services. It extends the
prompting service to support a hosting method that implements
stream-based prompting, allowing other methods to avoid streaming
requests and responses. This commit only converts service code - command
implementations have not yet been converted.

Signed-off-by: Jacob Howard <jacob@havoc.io>