Added conflict and problem truncation in list.
We've had several cases where some underlying issue (e.g. permissions,
different base paths, etc.) generates significant numbers of conflicts
and/or problems (sometimes in the thousands or even tens of thousands).
This is problematic because gRPC has a message size limit (which we've
already increased to 25 MB) that can prevent list operations from
working and thus prevent any insight into the underlying issue. It's
also problematic because it could (in theory) be a performance issue for
the daemon if there were frequent list operations or a large number of
concurrent list operations for such a session.

To remedy this, this commit introduces conflict and problem truncation.
This occurs at the service level and uses fixed maximum lenghts. These
lengths need to be sufficient to diagnose underlying issues, but not so
large that they cause sessions to exceed the gRPC message size limit
(not to mention making list output unreadable).

Signed-off-by: Jacob Howard <jacob@havoc.io>