Disabled support for netbsd/arm64 and openbsd/arm64.
While Go will build for these platforms, neither currently has a builder
and netbsd/arm64 is apparently broken. It looks like the Go 1.13 release
notes were updated to remove the announcement of these ports. Support
for netbsd/arm64 is tracked in golang/go#34036 and support for
openbsd/arm64 is tracked in golang/go#34035.

Signed-off-by: Jacob Howard <jacob@havoc.io>