winfsnotify: removeWatch: fix handle leak
When removing a watch to a directory we open a handle to it so we can
query the file index and volume id. Previously we then leaked the
handle. On some machines this would prevent a `rmdir dir ; mkdir dir`
from succeeding -- it would fail with an ACCESS_DENIED since the old
directory still exists because there is a handle to it.

We fix the leak by closing the handle as soon as we have used it to
query the information we need.

Signed-off-by: David Scott <dave.scott@docker.com>