Added resume operation to reconcileSessions.
This commit adds a session resume operation (for both forwarding and
synchronization sessions) to the reconcileSessions function. This is
essential for the case where sessions were paused using a stop
operation. It's also helpful in cases where the Mutagen service is being
started from a stopped state (e.g. due to a system shutdown) and the
sessions are in a reconnect loop waiting to reconnect.

Signed-off-by: Jacob Howard <jacob@havoc.io>