Switched compose pause/unpause commands to passthroughs.
Docker's paused states don't match up well with Mutagen's paused states,
which are more akin to start/stop states. With this commit, we switch
the pause/unpause commands to simple passthroughs, effectively disabling
pausing/unpausing of the Mutagen service (though still allowing it to be
stopped/started).

This change was motivated by the fact that an interrupt signal causes
the up command to stop containers, which clears their paused status, but
the Mutagen service runs independently from other services, so it was
very easy to get into a state where the Mutagen service was paused but
all others were not.

Signed-off-by: Jacob Howard <jacob@havoc.io>