Switched transport process termination to use SIGTERM on POSIX.
We were using SIGKILL for transport process termination, but this led to
zombie processes in certain cases (e.g. in issue #223). This commit
switches to using SIGTERM, which is not only more idiomatic, but also
doesn't have this problem. Of course, SIGTERM has some downsides, namely
that transport processes can ignore it and may not forward it properly,
but that doesn't really happen with our transports (SSH and Docker), and
even if it did we would still just block in a wait call until someone
kill -9'd the offending process. The only reason we were using SIGKILL
to begin with was because it was offered by a convenience method.

On Windows, we stick with os.Process.Kill, but this isn't sending a
signal, it's using TerminateProcess.

Signed-off-by: Jacob Howard <jacob@havoc.io>