Transitioned from Travis CI to GitHub Actions.
Travis has served Mutagen well, but the new usage quotas and pricing are
unfortunately unsustainable for the project. Consequently, this commit
coverts the project to use GitHub Actions for CI and release builds.
While this allows for better workflow structuring and integration with
GitHub, this commit aims to minimize dependencies on a specific CI
provider so that any future transitions will be less difficult. The
majority of the CI logic is now stored in modularized scripts, with the
CI platform only controlling toolchain/dependency setup and script
execution. The ideal balance of control between CI system and CI scripts
is somewhat subjective (e.g. one could have a primary run.sh script that
takes care of everything), so this commit opts for modular units of
execution corresponding to setup, testing, and building. This allows for
clearer control of the workflow in the CI configuration without becoming
overly reliant on any particular CI platform.

As part of this transition, the MUTAGEN_TEST_FAT32_ROOT environment
variable is now set for Windows tests. It previously wasn't, meaning the
virtual FAT32 volume created for tests on Windows wasn't being used at
all. Of course, there's probably not a huge amount of value in these
tests given how executability probing works on Windows, but it's good to
have for the sake of completeness.

Signed-off-by: Jacob Howard <jacob@havoc.io>