Combined CI and release workflows.
This commit folds the release workflow into the CI workflow and simply
makes the artifact deployment step conditional. Workflow artifacts are
used to pass bundles between build stages, with a 48 hour retention
period so that we can debug any issues without using an inordinate
amount of storage space.

This approach has a few important advantages. First, it allows us to
gate release artifact upload on all CI tests passing. Second, it
provides a mechanism for us to build Windows installers as part of the
CI pipeline. Finally, it allows us to avoid an additional release build
and ensures that the artifacts being released come from the same code
that was just tested.

Signed-off-by: Jacob Howard <jacob@havoc.io>