Expanded EINTR handling.
The standard library and runtime in Go 1.16 have further expanded EINTR
handling (to almost every system call), so this commit updates Mutagen
to follow suit.

This commit doesn't update EINTR handling on our vendored inotify code
because we'll be better off waiting for that to happen upstream (or,
ideally, replacing inotify with fanotify to the maximum extent
possible).

Signed-off-by: Jacob Howard <jacob@havoc.io>