Added buffering to error channel in Windows watcher.
This commit adds buffering to the watcher error channel on Windows,
which is necessary to fix the testing code that previously called
T.Fatalf from a background Goroutine. Adding buffering to this channel
allows us to simply check for errors at the end of the test. This code
should probably be rewritten at some point, but this fix should be
sufficient for the time being.

Signed-off-by: Jacob Howard <jacob@havoc.io>