Added watch completion wait to TestRecursiveWatchCycle.
This commit modifies TestRecursiveWatchCycle to wait until the
WatchRecursive Goroutine has terminated. This is necessary to avoid
flakiness when using testing.T.TempDir, which enforces successful
temporary directory removal (which can fail if any file descriptors or
handles are open on the directory).

This is part of the test flakiness fixes for f3dde29.

Signed-off-by: Jacob Howard <jacob@havoc.io>