Modified synchronization endpoint shutdown to wait on worker Goroutines.
This commit modifies synchronization endpoints to wait for background
worker Goroutine completion in the Shutdown method. This is necessary to
avoid flakiness when using testing.T.TempDir, which enforces successful
temporary directory removal (which can fail if any file descriptors or
handles are open on the directory). This isn't necessary for normal
Mutagen usage, only for testing, though it is probably best practice for
normal usage as well. This will also work for remote endpoints (at least
in testing) where the connection closure on shutdown will wait for
process termination. The one exception to that is the netpipe-based
protocol that we use in integration tests, but this commit also adds a
small wrapper in that case that will wait for full endpoint termination.

This is part of the test flakiness fixes for f3dde29.

Signed-off-by: Jacob Howard <jacob@havoc.io>