Enforced Windows symbolic link detection consistency with Go.
We were previously using an incorrect (or at least not completely
correct) symbolic link detection mechanism that differed from that of
the Go standard library. On Windows, where we use the os package to
drive some of our filesystem implementation, this led to spurious
symbolic link reports. This commit unifies Mutagen's symbolic link
definition with that of the Go standard library.

Note that the Go standard library still has an open issue regarding
symbolic link classification with WSL symbolic links (golang/go#42184),
so we inherit that for now, but we should be able to adjust our
definition once that issue is resolved.

Fixes #248.

Signed-off-by: Jacob Howard <jacob@havoc.io>