Removed EndpointOptions infrastructure.
EndpointOptions were originally designed for remote overrides of session
configuration elements, ideally for integration with non-agent-based
endpoints, but the idea never came to fruition, and the code wasn't
useful as the basis of any other features, so it was time to let it go.
It's always here if we need it.

Signed-off-by: Jacob Howard <jacob@havoc.io>