Downgraded to github.com/golang/protobuf v1.5.1.
Version 1.5.2 of github.com/golang/protobuf deprecates the module, but
we're not in a place to transition fully to the new Protocol Buffers
implementation because we're still using gRPC, the code generator from
the old module, and the encoding buffer implementation from the old
module. As such, it's best to stay on the pre-deprecated version so that
the Go toolchain doesn't start complaining about deprecation at an
inconvenient time.

Signed-off-by: Jacob Howard <jacob@havoc.io>