Removed Codecov uploading.
This commit removes support for Codecov. Coverage reports are still
generated, but we'll need to find a different way to combine and render
them. This is primarily in response to the Codecov security incident
announced on April 15, 2021. Mutagen does not appear to have been
affected by this incident, but the delay in reporting it, combined with
the lack of details in the report, are concerning.

Signed-off-by: Jacob Howard <jacob@havoc.io>