Fixed issue with staging resumption.
This commit fixes a problem with the stager, allowing it to use
directories that already exist. This was preventing staging resumption.
It seems to have missed detection/reporting due to the fact that the
staging would eventually stabilize on the next synchronization cycle,
but it ended up causing a wasted transmission of unstaged files and
(potentially) a whole lot of transition problems.

Signed-off-by: Jacob Howard <jacob@mutagen.io>