Removed YAML version information in Mutagen Compose.
Now that we're requiring docker-compose v1.28.0+, we no longer need to
propagate version information in our generated Compose YAML files,
because it's no longer required as of v1.27.0. We could continue to
support both presence and absence via an omitempty annotation on the
YAML field tag, but it's better just to remove the code.

Signed-off-by: Jacob Howard <jacob@mutagen.io>