Added hack to work around absence of .env files.
This commit adds a hackish fix for mutagen-io/mutagen#265. The issue
arose due to docker/compose@e0edc90 and docker/compose@4a26d95, which
subtly changed project resolution behavior (arguably in a way that isn't
backward-compatible (see, e.g., docker/compose#8264)). Essentially, the
project configuration files, which enter into the determination of the
project directory, are now actually computed twice: once before dotenv
loading and once after, and they factor in to the resolution of the
dotenv file path itself. This, somewhat indirectly, changed the
semantics of the --env-file flag in a way that was incompatible with how
Mutagen was using it.

At this point, the resolution behavior is just to difficult to emulate
100% faithfully. Since Mutagen's Compose functionality will soon be
replaced by something better, this commit adds a short hack that fixes
the issue while maintaining emulation fidelity in all but extreme edge
cases involving reevaluation of COMPOSE_FILES after dotenv loading. Most
people will never hit these cases.

If Compose emulation does live on, we can reevaluate this approach, but
we'd likely switch to supporting the Go-based Compose implementation
anyway, in which case we'll likely just use the underlying compose-go
implementation.

Fixes #265.

Signed-off-by: Jacob Howard <jacob@mutagen.io>