Added daemon termination logging.
This commit adds logging of daemon termination errors. It doesn't log
daemon startup errors since (a) doing so would be tedious and duplicate
a lot of code, (b) it's not possible for certain errors (such as lock
acquisition and log opening), and (c) debugging startup errors is more
easily accomplished using mutagen daemon run directly. In contrast,
termination (especially due to errors in the API server) is worth
logging so that debugging doesn't require keeping the Mutagen daemon
running in the foreground.

Signal-based termination has also been reclassified as a non-error.

Signed-off-by: Jacob Howard <jacob@mutagen.io>