Added support for TCP listening in the daemon.
This commit is the first step in transitioning the daemon away from
Unix Domain Sockets, Windows Named Pipes, and gRPC towards TCP, HTTP,
and REST. It adds a TCP listener to the daemon and a token generation
mechanism that clients can use to authenticate. No serving is currently
implemented, and the existing IPC/gRPC infrastructure is left in place
so that we can migrate services individually.

Signed-off-by: Jacob Howard <jacob@mutagen.io>