Switched daemon service to RESTful API.
This commit is the first in a series that will convert the daemon API to
a RESTful design using JSON. It adds an HTTP server to the daemon TCP
port and API endpoints for the daemon service. It also removes the old
daemon gRPC service, which required disabling some functionality until
this work is complete, though that functionality can currently be
accessed via manual HTTP requests.

Because the daemon gRPC service was previously responsible for
housekeeping, some refactoring/restructuring of the housekeeping
infrastructure was also required, though the result is better than what
existed previously.

Signed-off-by: Jacob Howard <jacob@mutagen.io>