Added support for an "agent" build tag to the build script.
This commit updates the build script to set an "agent" build tag when
building agent binaries. We were previously somewhat opposed to this
idea because we wanted the build script to be just a minimal wrapper
around go build, though (with our use of -ldflags and -trimpath) that
ship has sailed. Using an "agent" build tag will allow us to more easily
exclude code from the Go linker's field of vision when building agent
binaries, though (as with the -ldflags and -trimpath flags) it won't
prevent us from building the agent with a simple go build command (doing
so simply won't produce as size-optimized a build as the build script).

Being able to use this tag will be critical as we move HTTP-based
service definitions into their respective packages (because the net/http
package and its dependencies add ~10 MB to binaries).

Signed-off-by: Jacob Howard <jacob@mutagen.io>