Added sidecar image build to GitHub Actions.
This commit updates the GitHub Actions CI workflow to build the Mutagen
sidecar image and (if performing a release) push it to Docker Hub.

It also updates the Compose wrapper to pull a sidecar image version
corresponding to the exact Mutagen version tag.

Most importantly, this commit adds multi-architecture builds of the
sidecar image.

Updates mutagen-io/mutagen#276 and mutagen-io/sidecar#3.

Signed-off-by: Jacob Howard <jacob@mutagen.io>