Disabled S390x sidecar image build.
It seems like there's a problematic interaction between QEMU,
S390x (and RISC-V), and TLS (at least with ECDSA, but possibly other
encryption schemes as well). It has something to do with the emulation
of certain instructions used in hashing. For some background, see
golang/go#40949 and moby/moby#40240. This may be why the presence of the
linux/s390x image is somewhat inconsistent with golang Docker images.
Anyway, it's causing problems with the sidecar build, so this commit
disables S390x builds for the time being (which also allows us to use
more specific golang image tags). This shouldn't impact too many (if
any) users.

The only other option here would be to disable Go proxy verification for
S390x builds, which we might consider if S390x images become crucial,
but the same bug would probably plague the sidecar images themselves,
not just the build process.

Signed-off-by: Jacob Howard <jacob@mutagen.io>