Expanded sidecar CI job dependencies.
We only want to the sidecar to be pushed if we know that all build
products are ready to be shipped by the release job, so this commit
gates the sidecar build (and possibly push) on those other jobs
succeeding. We'll keep the release job gated on the sidecar job because
Mutagen will reference the sidecar image and we don't want a time window
(of any size) where that image hasn't been pushed.

This does have the unfortunate side-effect of blocking the start of the
sidecar build until later in the CI pipeline. If it becomes unbearable,
we can look into alternative options (e.g. building and pushing in
different jobs), but for now it shouldn't be a big issue since the
sidecar job only takes about 5 minutes to complete.

Signed-off-by: Jacob Howard <jacob@mutagen.io>