Updated issue template to be more free-form.
It turned out that there wasn't a huge amount of benefit from the rigid
issue template that we had previously, and it was more of a burden for
users. This commit restructures the issue template to be merely
informative and to request that users provide relevant information, with
some basic suggestions provided.

Signed-off-by: Jacob Howard <jacob@mutagen.io>