Added friendly names to some CI steps.
We don't want to add names to all CI steps, because doing so would make
the YAML file too verbose and obfuscate a lot of behavior. However,
there are a few steps where the label that's displayed automatically is
either derived from the first line of a script or is a run command whose
purpose isn't entirely obvious, and in those cases it's worth adding
human-friendly names.

Signed-off-by: Jacob Howard <jacob@mutagen.io>