Removed misleading comment.
While we technically could use scratch as a base image, it would
actually require a bit more work than just switching to scratch in the
Dockerfile. We'd need some sort of bootstrap executable to create the
/volumes directory and we'd need to create root's home directory and set
the HOME environment variable to support agent injection, etc. We could
build this functionality into the mutagen-sidecar executable, and
potentially even build a shell into that executable, but for the time
being it doesn't make sense to give up the functionality provided by
Alpine.

Signed-off-by: Jacob Howard <jacob@mutagen.io>