Replaced yamux with new multiplexing package.
This commit replaces yamux with the new multiplexing package, which
allows us to get rid of some of the internal wrapping packages that we
were using.

This commit also fixes a bug where agents doing reverse forwarding
wouldn't have accept operations preempted when the multiplexer and its
underlying connection were closed (resulting in zombie agents that held
on to listeners).

Signed-off-by: Jacob Howard <jacob@mutagen.io>