Updated github.com/bmatcuk/doubletar to v4.0.2.
This commit bumps github.com/bmatcuk/doubletar to v4.0.2. This is a
relatively large jump from v1.1.1, which we were holding on to until we
could verify that matching behavior hadn't changed substantially. After
reviewing the changes, it seems like all that's changed are edge case
fixes that we likely want the fixes for in any case. All of our tests
still pass, and the syntax still aligns with Git's ignore syntax.

Signed-off-by: Jacob Howard <jacob@mutagen.io>