Refactored code relying on errors.Wrap nil passthrough.
The github.com/pkg/errors package's Wrap function will return nil if the
error provided for wrapping is nil. For the purposes of automated
migration to fmt.Errorf-based wrapping, we don't want any code instances
that rely on this behavior. Fortunately we didn't use this functionality
very often, so there were only a few instances that needed to be
refactored. If we missed any, they'll show up as false-positive errors
when migrated to fmt.Errorf rather than false-negatives, so the
migration should be relatively safe.

Signed-off-by: Jacob Howard <jacob@mutagen.io>