Refactored remote endpoints to use io.ReadWriteCloser interface.
We previously used net.Conn as our interface of choice for remote
endpoints, just so that we could "guarantee" that the stream implemented
unblocking on Close, but realistically there was no way to guarantee
that via the interface alone, and implementing net.Conn cost a lot in
terms of additional boilerplate, most of which did absolutely nothing
(e.g. address and deadline implementations).

So, this commit switches to using io.ReadWriteCloser as the interface of
choice and simply documenting the required unblocking behavior.

Signed-off-by: Jacob Howard <jacob@mutagen.io>