Enabled preemption for flush requests.
Flush requests previously blocked all state change operations (such as
pause and terminate). While this made sense philosophically (since you
want a flush operation to be robust and reliable in some sense), it led
to a lot of cases of sessions that appeared to be hanging. The fact of
the matter is that if a pause or terminate command is run concurrently
with a flush command, even if started after the flush command, it could
still beat the flush command to connecting and operating on the session,
in which case the flush would fail. The rules are the same as with any
concurrent code: serialization is required for deterministic ordering.
Thus, it makes more sense for flush requests to be preemptable, and code
(typically code scripting Mutagen) that needs to ensure a flush
completes needs to hold off its pause and terminate requests until the
flush returns.

Signed-off-by: Jacob Howard <jacob@mutagen.io>